# Conflicts:
#	src/main/java/Database/MovieRepository/MovieRepo.java
#	src/main/java/Database/UserInfo/UserRepo.java

+++ src/main/java/Database/UserInfo/UserRepo.java
@@ -4,13 +4,12 @@ import com.mongodb.MongoClient;
 import com.mongodb.MongoClientSettings;
 import com.mongodb.MongoClientURI;
 import com.mongodb.client.MongoCollection;
+import com.mongodb.client.model.Filters;
 import org.bson.Document;
 import org.bson.codecs.configuration.CodecRegistries;
 import org.bson.codecs.configuration.CodecRegistry;
 import org.bson.codecs.pojo.ClassModel;
 import org.bson.codecs.pojo.PojoCodecProvider;
-
-import static com.mongodb.client.model.Filters.eq;
 
 public class UserRepo {
 
@@ -34,17 +33,21 @@ public class UserRepo {
     }
 
     public Chat getChat(Long chatId) {
-        return userRepo.find(eq("chatId", chatId)).first();
+        return userRepo.find(Filters.eq("chatId", chatId)).first();
     }
 
-    public void updateOrAddUser(long chatId) {
-        userRepo.deleteOne(new Document("chatId", chatId));
+    public void updateAndAddUser(long chatId) {
+        var chat = userRepo.find(Filters.eq("chatId", chatId)).first();
+        if (chat != null)
+            userRepo.deleteOne(new Document("chatId", chatId));
         userRepo.insertOne(new Chat(chatId, ChatState.START));
     }
 
     public void update(Chat user)
     {
-        userRepo.deleteOne(new Document("chatId", user.getChatId()));
+        var mongoUser = userRepo.find(Filters.eq("chatId", user.getChatId())).first();
+        if (mongoUser != null)
+            userRepo.deleteOne(new Document("chatId", user.getChatId()));
         userRepo.insertOne(user);
     }
 }
